tests: adding test to run spread tests in a nested machine
This test is gonna be used to run spread tests in a ubuntu core vm for
arch i386 and amd64.

The idea is to run nightly the master in order to detect any issue on
real ubuntu core OS, avoiding failures when it is tested on beta.

The change allows also to run setup commands in the machine where spread
tests are executed and also in the nested one in order to be able to run
scenarios like the beta refresh.